requirements.txt
pandas==1.1.5
xlrd==2.0.1
python-dotenv==0.20.0